Revert "Squashed commit of the following:"
This reverts commit 272e6f08c0f90e025def25459abb038c279aec13.

+++ pandatool/src/pstatserver/pStatThreadData.cxx
@@ -235,7 +235,7 @@ get_frame_number_after(double time, int start_at) const {
 
   while (end < time) {
     ++i;
-    if ((size_t)i >= _frames.size()) {
+    if (i >= _frames.size()) {
       break;
     }
     if (_frames[i] != nullptr) {
